refactor: add directory for user-auth-slice
