+++ src/pages/auth/authentication-page/authentication-page.jsx
@@ -8,7 +8,7 @@ import { Link, useNavigate } from 'react-router-dom';
 import { ButtonFieldset, EmailFieldset, PasswordFieldset } from '../../../components/UI/index';
 import style from './authentication-page.module.css';
 import oauthSignIn from '../../../utils/google-request';
-import { loginUser } from '../../../services/slice/user-auth-slice';
+import { loginUser } from '../../../services/slice/auth/user-auth-slice';
 import { GoogleBtn, Form } from '../../../components/UI/index';
 import { getLoginErrorText } from '../../../services/selectors/selectors';
