+++ src/pages/auth/google-auth-page/google-auth-page.jsx
@@ -3,7 +3,7 @@ import { useDispatch } from 'react-redux';
 import { useSearchParams, useNavigate } from 'react-router-dom';
 import Loader from '../../../components/UI/loader/loader';
 import style from './google-auth-page.module.css';
-import { fetchGoogleDate } from '../../../services/slice/user-auth-slice';
+import { fetchGoogleDate } from '../../../services/slice/auth/user-auth-slice';
 
 function GoogleAuthPage() {
   const dispatch = useDispatch();
