+++ src/pages/auth/registration-page/registration-page.jsx
@@ -14,7 +14,7 @@ import {
   TextFieldset,
 } from '../../../components/UI/index';
 import style from './registration-page.module.css';
-import { registerUser, setEmail } from '../../../services/slice/user-auth-slice';
+import { registerUser, setEmail } from '../../../services/slice/auth/user-auth-slice';
 import oauthSignIn from '../../../utils/google-request';
 
 function RegistrationPage() {
