+++ src/pages/auth/verify-email-page/verify-email-page.jsx
@@ -3,7 +3,7 @@ import { useDispatch } from 'react-redux';
 import { useParams, useNavigate } from 'react-router-dom';
 import Loader from '../../../components/UI/loader/loader';
 import style from './verify-email-page.module.css';
-import { confirmUserEmail } from '../../../services/slice/user-auth-slice';
+import { confirmUserEmail } from '../../../services/slice/auth/user-auth-slice';
 
 function VerifyEmailPage() {
   const dispatch = useDispatch();
