+++ src/services/slice/auth/user-auth-slice.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-param-reassign */
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
-import api from '../../api/Api';
-import token from '../../utils/token';
+import api from '../../../api/Api';
+import token from '../../../utils/token';
 
 export const fetchGoogleDate = createAsyncThunk('userAuth/fetchGoogleDate', (googleCode) =>
   api.googleAuthRequest(googleCode).then((res) => {
